feat(comp): clean up - move code in own file
Refs: #38
