+++ components/dice/command.go
@@ -2,8 +2,6 @@
 package dice
 
 import (
-	"fmt"
-	"strconv"
 	"github.com/bwmarrin/discordgo"
 )
 
@@ -13,9 +11,7 @@ func handleDice(s *discordgo.Session, i *discordgo.InteractionCreate) {
 	d := getIntOption(i, "die-sites-number", 6)
 
 	r := rollDice(n, d)
-	s := arrayIntToArrayString(r)
-	a := createAnswer(n, d, s)
-	sendAnswer(a)
+	sendAnswerToUser(n, d, r)
 }
 
 // create a map and insert the command options
@@ -37,51 +33,4 @@ func getIntOption(options map[string]*discordgo.ApplicationCommandInteractionDat
 	}
 
 	return 0
-}
-
-// Create the answer to send
-func createAnswer(n int, d int, rolledDice []int) string {
-	answer := fmt.Sprintf("You rolled %d d%d, The Results are:\n", n, d)
-	
-	answer += implode(", ", rolledDice)
-
-	return answer
-}
-
-// Send the Answer
-func sendAnswer(answerText string) {
-	_ = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-		Type: discordgo.InteractionResponseChannelMessageWithSource,
-		Data: &discordgo.InteractionResponseData{
-			Content: answerText,
-		}
-	})
-}
-
-// Create one string with the array of string seperated with seperator (s)
-func implode(s string, array []string) string {
-	first := true
-	r := ""
-
-	for _, a := range array {
-		if (first) {
-			first = false
-			r += a
-		} else {
-			r += s + a
-		}
-	}
-
-	return r
-}
-
-// Convert an array of int to an array of string
-func arrayIntToArrayString(ints []int) []string {
-	strings := []string {}
-
-	for k, i := range ints {
-		strings[k] = strconf.Itoa(i)
-	}
-
-	return strings
 }
